# examples/demo-backend/openapi/complex.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/testing/scenarios": {
      "get": {
        "operationId": "ComplexTesting_scenarios",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenariosResponseDto"
                }
              }
            }
          }
        },
        "summary": "List deterministic X-Demo-Scenario values",
        "tags": [
          "Testing"
        ]
      }
    },
    "/api/v1/testing/reset": {
      "post": {
        "operationId": "ComplexTesting_reset",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestingActionResponseDto"
                }
              }
            }
          }
        },
        "summary": "Reset all Complex API data and active sessions",
        "tags": [
          "Testing"
        ]
      }
    },
    "/api/v1/testing/seed/{preset}": {
      "post": {
        "operationId": "ComplexTesting_seed",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "preset",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "small",
                "large"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestingActionResponseDto"
                }
              }
            }
          }
        },
        "summary": "Select a small or large deterministic dataset",
        "tags": [
          "Testing"
        ]
      }
    },
    "/api/v1/testing/session/expire": {
      "post": {
        "operationId": "ComplexTesting_expireSession",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestingActionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "csrf": []
          },
          {
            "cookieSession": []
          }
        ],
        "summary": "Expire the current cookie session after this response",
        "tags": [
          "Testing"
        ]
      }
    },
    "/api/v1/testing/users/{userId}/role": {
      "post": {
        "operationId": "ComplexTesting_changeRole",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeComplexRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplexUserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Change a role while sessions remain active",
        "tags": [
          "Testing"
        ]
      }
    }
  },
  "info": {
    "title": "Demo Complex API",
    "description": "Cookie-session, multitenant and realtime API for large frontend applications.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "demo_session"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Required for authenticated mutations."
      }
    },
    "schemas": {
...
      "ComplexUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "complex-user-admin"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "admin@complex.demo"
          },
          "name": {
            "type": "string",
            "example": "Complex Admin"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "manager",
              "support",
              "viewer"
            ]
          },
          "avatarUrl": {
            "type": "object",
            "nullable": true,
            "example": "https://i.pravatar.cc/160?img=20"
          },
          "organizationIds": {
            "example": [
              "org-acme",
              "org-globex"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "avatarUrl",
          "organizationIds"
        ]
      },
...
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "email"
          },
          "message": {
            "type": "string",
            "example": "must be an email"
          },
          "code": {
            "type": "string",
            "example": "isEmail"
          }
        },
        "required": [
          "message"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "code": {
            "type": "string",
            "example": "PRODUCT_NOT_FOUND"
          },
          "message": {
            "type": "string",
            "example": "Product not found"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetailDto"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-30T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/products/product-404"
          },
          "requestId": {
            "type": "string",
            "example": "req-5c9f7a3d"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "details",
          "timestamp",
          "path",
          "requestId"
        ]
      },
      "ComplexUserResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ComplexUserDto"
          }
        },
        "required": [
          "data"
        ]
      },
...
      "ScenarioDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "slow"
          },
          "description": {
            "type": "string",
            "example": "Delays the response to exercise loading and cancellation states."
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
      "ScenariosResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScenarioDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "TestingActionDataDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "State reset to the default deterministic seed."
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
      "TestingActionResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TestingActionDataDto"
          }
        },
        "required": [
          "data"
        ]
      },
      "ChangeComplexRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "manager",
              "support",
              "viewer"
            ],
            "example": "viewer"
          }
        },
        "required": [
          "role"
        ]
      }
    }
  }
}